refactor: 연결 부분을 jdbc로 분리
