+++ src/main/java/db/MovementDao.java
@@ -1,35 +1,17 @@
 package db;
 
 import domain.board.position.Position;
-import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
 public final class MovementDao {
 
-    private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
-    private static final String DATABASE = "chess"; // MySQL DATABASE 이름
-    private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
-    private static final String USERNAME = "root"; //  MySQL 서버 아이디
-    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
-
-    public Connection getConnection() {
-        // 드라이버 연결
-        try {
-            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
-        } catch (final SQLException e) {
-            System.err.println("DB 연결 오류:" + e.getMessage());
-            e.printStackTrace();
-            return null;
-        }
-
-    }
+    JdbcConnector jdbcConnector = new JdbcConnector();
 
     public void createMovement(final Movement movement) {
         final var query = "INSERT INTO movement (source,  target, shape, color) values(?, ?, ?, ?)";
-        try (final var connection = getConnection();
+        try (final var connection = jdbcConnector.getConnection();
              final var preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setString(1, movement.source().toString());
             preparedStatement.setString(2, movement.target().toString());
@@ -43,7 +25,7 @@ public final class MovementDao {
 
     public List<Movement> findAll() {
         final var query = "SELECT * FROM movement";
-        try (final var connection = getConnection();
+        try (final var connection = jdbcConnector.getConnection();
              final var preparedStatement = connection.prepareStatement(query)) {
             final var resultSet = preparedStatement.executeQuery();
             final List<Movement> movements = new ArrayList<>();
@@ -63,7 +45,7 @@ public final class MovementDao {
 
     public void deleteAll() {
         final var query = "DELETE from movement";
-        try (final var connection = getConnection();
+        try (final var connection = jdbcConnector.getConnection();
              final var preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.executeUpdate();
         } catch (final SQLException e) {
