+++ src/test/java/db/JdbcConnectorTest.java
@@ -1,0 +1,15 @@
+package db;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class JdbcConnectorTest {
+
+    @Test
+    @DisplayName("Jdbc가 연결되었는지 확인한다")
+    void getConnection() {
+        assertThatCode(() -> new JdbcConnector().getConnection()).doesNotThrowAnyException();
+    }
+}
